added ability to reply with mention tags

+++ src/components/Comments.jsx
@@ -3,11 +3,12 @@ import "../styles/Comments.css";
 import useFetch from "../utilities/useFetch";
 import Editor from "./Editor";
 
-function Comments({ id }) {
+function Comments({ parent }) {
   const [comments, setComments] = useState([]);
   const [showEditor, setShowEditor] = useState(false);
   const [showControls, setShowControls] = useState(null);
-  const { data: fetchData, status: fetchStatus } = useFetch(`${import.meta.env.VITE_API_URL}/messages/${id}`, false);  
+  const [mention, setMention] = useState(null);
+  const { data: fetchData, status: fetchStatus } = useFetch(`${import.meta.env.VITE_API_URL}/messages/${parent}`, false);  
 
   const blogID = new URLSearchParams(window.location.search).get("blog");
   
@@ -20,13 +21,21 @@ function Comments({ id }) {
   if (fetchStatus == "error")
     return (<div className="error comments"> <div>&#x2716;</div> Oops! Something went wrong. </div>);
 
-  const handleDisplayControls = (commentId) => {
-    setShowControls(showControls === commentId ? null : commentId);
+  let handleDisplayControls = (id) => {
+    setShowControls(showControls === id ? null : id);
   };
 
-  const handleDelete = async (commentId) => {
+  let handleMention = (user) => {
+    setMention(user);
+    setShowEditor(true);
+  };
+
+  let handleDelete = async (id) => {
+    const confirmDelete = window.confirm('Delete this comment?');
+    if (!confirmDelete) return;
+
     try {
-      const response = await fetch(`${import.meta.env.VITE_API_URL}/messages/${commentId}`, {
+      const response = await fetch(`${import.meta.env.VITE_API_URL}/messages/${id}`, {
         method: 'DELETE',
         headers: { 'Content-Type': 'application/json' },
         body: JSON.stringify({
@@ -35,7 +44,7 @@ function Comments({ id }) {
         }),
       });
       if (response.ok) {
-        setComments(comments.filter(comment => comment.id !== commentId));
+        setComments(comments.filter(comment => comment.id !== id));
       } else {
         console.error('Failed to delete the comment');
       }
@@ -46,7 +55,7 @@ function Comments({ id }) {
 
   return (
     <div className="comments">
-      { blogID == id && <Editor {...{id, setComments, setShowEditor}} /> }
+      { blogID == parent && <Editor {...{parent, setComments, setShowEditor}} /> }
       {
         comments.map((comment) =>
           <div className="comments-list" key={comment.id} title={new Date(comment.date).toLocaleString().concat(comment.status == "edited" ? " (edited)" : "")}>
@@ -56,21 +65,21 @@ function Comments({ id }) {
                 <span className="comment-controls" onClick={() => handleDisplayControls(comment.id)}> ...
                   {showControls === comment.id && (
                     <ul className="comment-controls-list">
-                      <li>Reply</li>
-                      <li>Modify</li>
-                      <li onClick={() => handleDelete(comment.id)}>Delete</li>
+                      <li onClick={() => handleMention(comment.user)}>💬 Reply</li>
+                      <li>✏️ Modify</li>
+                      <li onClick={() => handleDelete(comment.id)}>🗑️ Delete</li>
                     </ul>
                   )}
                 </span>
               </div>
               <div className="comment-text">{comment.content}</div>
             </div>
-            { blogID == id && <Comments id={comment.id} /> }
+            { blogID == parent && <Comments parent={comment.id} /> }
           </div>
         )
       }
-      { showEditor && blogID != id && <Editor {...{id, setComments, setShowEditor}} /> }
-      { blogID != id && !showEditor && <button onClick={() => setShowEditor(!showEditor)}> Reply </button> }
+      { showEditor && blogID != parent && <Editor {...{parent, mention, setComments, setShowEditor}} /> }
+      { blogID != parent && !showEditor && <button onClick={() => setShowEditor(!showEditor)}> Reply </button> }
     </div>
   );
 }
